pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mdc-encyclopedia"
version = "0.1.0"
description = "Miami-Dade County Open Data Encyclopedia"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Diego Abreu" },
]
dependencies = [
    "click>=8.0",
    "rich-click>=1.8",
    "rich>=13.0",
    "httpx>=0.28",
    "tenacity>=9.0",
    "anthropic>=0.80",
    "python-dotenv>=1.0",
]

[project.scripts]
mdc-encyclopedia = "mdc_encyclopedia.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/mdc_encyclopedia"]
